refactor : FirebaseTokenFilter 추상화 삭제, 필터 메소드 내용 일부 수정
- Prod, Inter FirebaseTokenFilter 삭제, 부모 클래스 수정
- FirebaseTokenFilter클래스로 통합
- 필터 메소드 내용 중 loadUserByUsername()으로 가져온 user에 uid 넣어주도록 수정
